Refactored login system
