+++ src/main/java/controllers/AdministratorTransaction.java
@@ -1,4 +1,4 @@
-package transactions;
+package controllers;
 
 import database.DatabaseAccessBoundary;
 import users.Administrator;
