+++ src/main/java/controllers/TransactionFactory.java
@@ -1,4 +1,4 @@
-package transactions;
+package controllers;
 
 import database.DatabaseAccessBoundary;
 import uses.InventorySystem;
