+++ src/main/java/database/data/Fruits_and_Vegetables.json
@@ -1,1 +1,1 @@
-[{"unit":"1 each","quantity":263,"price":0.85,"name":"Cucumber"},{"unit":"1 bunch","quantity":1000,"price":0.25,"name":"Green Onion"},{"unit":"1kg","quantity":1000,"price":1.0,"name":"Yellow Onions"},{"unit":"1kg","quantity":1000,"price":3.5,"name":"Red Onion"},{"unit":"1 each","quantity":1000,"price":0.5,"name":"Sweets Peppers"},{"unit":"1kg","quantity":1000,"price":2.0,"name":"Sweet Potato"},{"unit":"1kg","quantity":1000,"price":4.0,"name":"Ginger"},{"unit":"1kg","quantity":1000,"price":0.9,"name":"Green Cabbage"},{"unit":"100g","quantity":1000,"price":1.0,"name":"Garlic"},{"unit":"100g","quantity":1000,"price":0.3,"name":"Mini Carrots"},{"unit":"1kg","quantity":1000,"price":0.9,"name":"Green Cabbage"},{"unit":"1kg","quantity":1000,"price":2.0,"name":"Carrots"},{"unit":"100g","quantity":1000,"price":0.5,"name":"Mushrooms"},{"unit":"1 bunch","quantity":1000,"price":1.0,"name":"Banana"},{"unit":"1 box","quantity":1000,"price":3.0,"name":"Strawberries"},{"unit":"1 box","quantity":1000,"price":3.0,"name":"Kiwis"},{"unit":"1kg","quantity":1000,"price":3.0,"name":"Red Grapes"},{"unit":"1kg","quantity":1000,"price":3.0,"name":"Green Grapes"},{"unit":"1 each","quantity":1000,"price":2.0,"name":"Cantaloupe"},{"unit":"1 each","quantity":1000,"price":1.99,"name":"Peach"},{"unit":"1 each","quantity":1000,"price":1.0,"name":"Pomegranate"},{"unit":"1 each","quantity":1364,"price":0.99,"name":"Apple"},{"unit":"1 each","quantity":1000,"price":1.25,"name":"Pineapple"}]
+[{"unit":"1 each","quantity":243,"price":0.85,"name":"Cucumber"},{"unit":"1 bunch","quantity":1000,"price":0.25,"name":"Green Onion"},{"unit":"1kg","quantity":1000,"price":1.0,"name":"Yellow Onions"},{"unit":"1kg","quantity":1000,"price":3.5,"name":"Red Onion"},{"unit":"1 each","quantity":1000,"price":0.5,"name":"Sweets Peppers"},{"unit":"1kg","quantity":1000,"price":2.0,"name":"Sweet Potato"},{"unit":"1kg","quantity":1000,"price":4.0,"name":"Ginger"},{"unit":"1kg","quantity":1000,"price":0.9,"name":"Green Cabbage"},{"unit":"100g","quantity":1000,"price":1.0,"name":"Garlic"},{"unit":"100g","quantity":1000,"price":0.3,"name":"Mini Carrots"},{"unit":"1kg","quantity":1000,"price":0.9,"name":"Green Cabbage"},{"unit":"1kg","quantity":1000,"price":2.0,"name":"Carrots"},{"unit":"100g","quantity":1000,"price":0.5,"name":"Mushrooms"},{"unit":"1 bunch","quantity":1000,"price":1.0,"name":"Banana"},{"unit":"1 box","quantity":1000,"price":3.0,"name":"Strawberries"},{"unit":"1 box","quantity":1000,"price":3.0,"name":"Kiwis"},{"unit":"1kg","quantity":1000,"price":3.0,"name":"Red Grapes"},{"unit":"1kg","quantity":1000,"price":3.0,"name":"Green Grapes"},{"unit":"1 each","quantity":1000,"price":2.0,"name":"Cantaloupe"},{"unit":"1 each","quantity":1000,"price":1.99,"name":"Peach"},{"unit":"1 each","quantity":1000,"price":1.0,"name":"Pomegranate"},{"unit":"1 each","quantity":1364,"price":0.99,"name":"Apple"},{"unit":"1 each","quantity":1000,"price":1.25,"name":"Pineapple"}]
