+++ src/main/java/database/data/Household_Supplies.json
@@ -1,1 +1,1 @@
-[{"unit":"1 each","quantity":100,"price":5.99,"name":"Paper Towel - 6 Packs"},{"unit":"1 each","quantity":50,"price":5.0,"name":"Hypoallergenic - 12 Packs"},{"unit":"1 each","quantity":50,"price":4.5,"name":"Tissue - 6 Packs"},{"unit":"1 each","quantity":50,"price":2.0,"name":"Sandwich Bags"},{"unit":"1 each","quantity":50,"price":5.0,"name":"Yard Waste Bags - 10 Pack"},{"unit":"1 each","quantity":50,"price":3.0,"name":"Aluminum Foil 50 feet"},{"unit":"1 each","quantity":50,"price":4.5,"name":"Garbage Bags - 40 Packs"},{"unit":"1 each","quantity":50,"price":1.3,"name":"Dishwashing Liquid"},{"unit":"1 each","quantity":1000,"price":5.99,"name":"Broom"}]
+[{"unit":"1 each","quantity":750,"price":6.99,"name":"Paper Towel - 6 Packs"},{"unit":"1 each","quantity":50,"price":5.0,"name":"Hypoallergenic - 12 Packs"},{"unit":"1 each","quantity":50,"price":4.5,"name":"Tissue - 6 Packs"},{"unit":"1 each","quantity":50,"price":2.0,"name":"Sandwich Bags"},{"unit":"1 each","quantity":50,"price":5.0,"name":"Yard Waste Bags - 10 Pack"},{"unit":"1 each","quantity":50,"price":3.0,"name":"Aluminum Foil 50 feet"},{"unit":"1 each","quantity":50,"price":4.5,"name":"Garbage Bags - 40 Packs"},{"unit":"1 each","quantity":50,"price":1.3,"name":"Dishwashing Liquid"},{"unit":"1 each","quantity":1000,"price":5.99,"name":"Broom"}]
