+++ src/main/java/database/data/New_Products.json
@@ -1,1 +1,1 @@
-[{"unit":"1 each","quantity":10,"price":2.99,"name":"Vanilla Ice Cream"},{"unit":"1 each","quantity":57,"price":3.99,"name":"Chocolate Ice Cream"},{"unit":"1 each","quantity":152,"price":2.79,"name":"Strawberry Ice Cream"},{"unit":"1 each","quantity":100,"price":4.99,"name":"Mint Ice Cream"},{"unit":"1 each","quantity":10,"price":2.99,"name":"Assorted Ice Cream"},{"unit":"1 each","quantity":100,"price":6.99,"name":"Frozen Cheese Pizza"}]
+[{"unit":"1 each","quantity":10,"price":2.99,"name":"Vanilla Ice Cream"},{"unit":"1 each","quantity":57,"price":3.99,"name":"Chocolate Ice Cream"},{"unit":"1 each","quantity":152,"price":2.79,"name":"Strawberry Ice Cream"},{"unit":"1 each","quantity":100,"price":4.99,"name":"Mint Ice Cream"},{"unit":"1 each","quantity":10,"price":2.99,"name":"Assorted Ice Cream"},{"unit":"1 each","quantity":100,"price":6.99,"name":"Frozen Cheese Pizza"},{"unit":"1 each","quantity":1200,"price":0.99,"name":"Lollipop"}]
