--- src/main/java/login/LoginUI.java
@@ -1,18 +1,0 @@
-package login;
-
-import java.util.Scanner;
-
-/**
- * A beautiful user interface for logging in.
- */
-public class LoginUI {
-
-    public void runLogin(LoginController controller) {
-        System.out.println("Username: ");
-        Scanner scanner = new Scanner(System.in);
-        String username = scanner.nextLine();
-        System.out.println("Password: ");
-        String password = scanner.nextLine();
-        controller.runLogin(username, password);
-    }
-}
