+++ src/main/java/ui/MenuUI.java
@@ -2,7 +2,7 @@ package ui;
 
 import database.DatabaseAccess;
 import prompts.PromptIterator;
-import transactions.TransactionSystem;
+import controllers.TransactionSystem;
 
 import java.io.File;
 import java.io.IOException;
