+++ src/test/java/TransactionSystemTest.java
@@ -1,6 +1,6 @@
 import database.DatabaseAccess;
 import org.junit.Test;
-import transactions.TransactionSystem;
+import controllers.TransactionSystem;
 
 
 import java.util.ArrayList;
